Move division to separate modules
